Modularize stopping body scrolling and Encapsulte closing modal with esc key
